# docs/swagger/openapi.yaml
openapi: "3.0.3"
info:
  title: Workflow Definition Service
  version: "1.0.0"
  description: >
    Manages workflow template lifecycle: create, draft, publish, archive, clone,
    promote, validate, export, and diff BPMN-based workflow definitions.
    All identity and tenant context is injected by the Envoy API Gateway.
servers:
  - url: http://localhost:8080
    description: Local dev server (matches cmd/server's actual route mounts)
  - url: https://api.internal/definition/v1
    description: Envoy API Gateway (production)
# All endpoints require a valid tenant context injected by the Envoy gateway.
# Mutating endpoints (POST/PUT/DELETE) additionally require x-tenant-roles.
security:
  - GatewayTenantContext: []
components:
  securitySchemes:
    GatewayTenantContext:
      type: apiKey
      in: header
      name: x-tenant-id
      description: >
        Tenant UUID injected by the Envoy ext_authz filter after JWT validation.
        Must be present on every request. Never supplied directly by the client
        in production — the gateway rejects any request without a valid JWT and
        sets this header from the token claims.

    AdminRoles:
      type: apiKey
      in: header
      name: x-tenant-roles
      description: >
        Roles claim injected by the gateway. Required for mutating operations.
        Contains a comma-separated list of roles (e.g. tenant_admin,member).
        Accepted values for mutation: tenant_admin, tenant_owner.
        Read-only endpoints (GET) do not require this header.

    InternalServiceAuth:
      type: apiKey
      in: header
      name: x-internal-token
      description: >
        Service-to-service credential for the internal route group (e.g. POST /internal/events).
        These routes are NOT exposed through the Envoy public gateway and do NOT carry gateway
        identity headers; they are restricted to in-mesh callers by Kubernetes NetworkPolicy and
        authenticated by this shared secret (or mTLS, depending on deployment). The tenant context
        for these routes is taken from the request payload, not from x-tenant-id.
  parameters:
    TenantId:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant UUID injected by Envoy.
    UserId:
      name: x-user-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Authenticated user UUID injected by Envoy.
    TenantRoles:
      name: x-tenant-roles
      in: header
      required: true
      schema:
        type: string
      description: Comma-separated tenant roles injected by Envoy.
    Departments:
      name: x-departments
      in: header
      required: false
      schema:
        type: string
      description: Comma-separated list of department memberships and role levels injected by Envoy.
    Plan:
      name: x-plan
      in: header
      required: false
      schema:
        type: string
      description: Tenant billing plan tier injected by Envoy.
    FeatureFlags:
      name: x-feature-flags
      in: header
      required: false
      schema:
        type: string
      description: Encoded set of effective feature flags injected by Envoy.
    WorkflowId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VersionId:
      name: version_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    VersionStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
...
    ValidationError:
      type: object
      properties:
        node_id: { type: string }
        error:   { type: string }
      example:
        node_id: Task_design_prep
        error: "userTask is missing zeebe:taskDefinition"
...
    StageDef:
      type: object
      required: [type, activity, role]
      properties:
        type:
          type: string
          description: >
            Stage type sourced from zeebe:taskDefinition on the BPMN element.
            Built-in values: prep, review, approve (userTask); send_task, receive_task (sendTask/receiveTask).
            Unknown types are forwarded as passthrough with engine_note set.
        activity:            { type: string }
        role:
          type: string
          description: "IAM role level name passed as '?level=<role>' to the Org & Membership eligibility endpoint at publish time. Sourced from zeebe:assignmentDefinition candidateGroups."
        default_assignees:
          type: array
          nullable: true
          items: { type: string, format: uuid }
          minItems: 1
          description: "Default assignee UUID(s). Sourced from zeebe:assignmentDefinition candidateUsers. Omitted when assignment is not set — optional for send_task/receive_task stage types."
        is_zeebe_user_task:
          type: boolean
          description: "true when the userTask carries a bare <zeebe:userTask/> marker extension element (no taskDefinition/assignmentDefinition required in that case). Omitted (false) otherwise."
        node_id:
          type: string
          description: "Stable BPMN element ID (e.g. 'Activity_0abc123'). Suitable for machine-addressable routing."
        due_date:
          type: string
          nullable: true
          description: "FEEL expression or ISO 8601 datetime for the task due date. Sourced from zeebe:taskSchedule dueDate. Absent when not set in BPMN."
        follow_up_date:
          type: string
          nullable: true
          description: "FEEL expression or ISO 8601 datetime for the task follow-up date. Sourced from zeebe:taskSchedule followUpDate. Absent when not set in BPMN."
        extras:
          type: object
          nullable: true
          additionalProperties: { type: string }
          description: "Forward of every zeebe:property from the BPMN task, unmodified — the compiler does not special-case any property name. Also carries 'message' for send_task/receive_task stage types."
        engine_note:
          type: string
          nullable: true
          description: "Advisory note set when stage type is not a registered engine class (IAM-owned passthrough). Absent for known types."
        boundary_timer:
          type: object
          nullable: true
          description: "SLA deadline from a timer boundary event on the task. Null when no boundary timer is attached."
          properties:
            duration:     { type: string, description: "ISO 8601 duration (e.g. 'PT48H', 'P3D')." }
            interrupting: { type: boolean, description: "true = task cancelled on fire; false = escalation runs concurrently." }
            target_dept:  { type: string, nullable: true, description: "Department key to route the escalation path to, if set." }
        boundary_message:
          allOf: [{ $ref: '#/components/schemas/MessagePath' }]
          nullable: true
          description: "Correlation info from a message boundary event on the task. Null when no message boundary is attached."
      example:
        type: prep
        activity: PrepActivity
        node_id: Activity_0abc123
        role: preparator
        default_assignees: [550e8400-e29b-41d4-a716-446655440000]
        extras: { requires_comment: "false" }
        boundary_timer: { duration: "PT24H", interrupting: false }
    DepartmentDef:
      type: object
      required: [id, label, stages]
      properties:
        id:
          type: string
          description: "Department identifier from the BPMN lane name. Sourced from the lane name attribute; format is owned by the Profile Service."
        label:  { type: string }
        ignore:
          type: boolean
          description: "true when the lane carries a <zeebe:property name=\"ignore\" value=\"true\"/> extension — marks an ignored-pool lane. Omitted (false) otherwise."
        props:
          type: object
          nullable: true
          additionalProperties: { type: string }
          description: "Arbitrary key/value pairs forwarded from <zeebe:properties> on the lane element."
        stages:
          type: array
          items: { $ref: '#/components/schemas/StageDef' }
      example:
        id: Tender
        label: Tender
        stages:
          - type: prep
            activity: PrepActivity
            role: bd-agent
            default_assignees: [018e1f2a-0000-7000-8000-000000000001]
            extras: { requires_comment: "false" }
          - type: review
            activity: ReviewActivity
            role: tender-manager
            default_assignees: [018e1f2a-0000-7000-8000-000000000002]
            extras: { requires_comment: "true" }
          - type: approve
            activity: ApproveActivity
            role: tender-director
            default_assignees: [018e1f2a-0000-7000-8000-000000000003]
            extras: { requires_comment: "true" }
    ExecutionStep:
      type: object
      properties:
        sequential:
          type: array
          items: { type: string }
        parallel:
          type: array
          description: "Each entry is a parallel branch with its own sequential execution steps."
          items:
            $ref: '#/components/schemas/ParallelBranch'
        exclusive:
          type: array
          items:
            type: object
            required: [condition_expression]
            properties:
              target:
                type: string
                description: >
                  Destination department for a forward branch. Omitted when the branch
                  terminates the workflow (see terminates) or on a revert branch.
              target_stage:  { type: string }
              target_node_id:
                type: string
                description: "Stable BPMN element ID of the forward target. Use this for machine routing instead of target_name."
              target_name:
                type: string
                description: >
                  BPMN element name of the target node. Present when target_stage is
                  "sub_workflow" (disambiguates multiple subworkflows in the same dept)
                  or empty (task has no zeebe:taskDefinition).
              condition_expression: { type: string }
              terminates:
                type: boolean
                description: "True when this forward branch leads directly to an end event."
              revert_to_dept:
                type: string
                description: >
                  Set on a guarded-loop revert branch (a back-edge from this exclusive gateway):
                  the department to revert to. Present instead of a populated target.
              revert_to_stage:
                type: string
                description: "Set with revert_to_dept: the stage level to revert to."
              revert_to_node_id:
                type: string
                description: "Stable BPMN element ID of the revert target. Use for machine routing."
              revert_to_name:
                type: string
                description: >
                  BPMN element name of the revert target. Present when revert_to_stage
                  is "sub_workflow" or empty, same rules as target_name.
        sub_workflow:
          $ref: '#/components/schemas/SubWorkflowStep'
          nullable: true
        call_pool:
          type: object
          nullable: true
          description: "Present when this step hands control to another compiled pool (Temporal child workflow). Only emitted by the main pool."
          required: [pool]
          properties:
            pool: { type: string, description: "Name of the target CompiledPlan." }
        io_mapping:
          type: object
          nullable: true
          description: "Variable input/output declarations for a callActivity step. Attached to the first step of a flattened callActivity. The Execution Service uses this to pass variables when creating child workflows. dept_id is compiler-internal and is NOT emitted here."
          properties:
            inputs:
              type: array
              items:
                type: object
                required: [source, target]
                properties:
                  source: { type: string, description: "FEEL expression or literal variable name from the caller's context." }
                  target: { type: string, description: "Variable name in the called process context." }
            outputs:
              type: array
              items:
                type: object
                required: [source, target]
                properties:
                  source: { type: string }
                  target: { type: string }
        extras:
          type: object
          nullable: true
          additionalProperties: { type: string }
          description: "Forward of unknown zeebe:properties from a callActivity element. Attached to the first step of a flattened callActivity."
        message_paths:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/MessagePath' }
          description: "Message boundary events attached to a callActivity, flattened into its first step."
      example:
        sequential: [design, planning, construction]
    ParallelBranch:
      type: object
      required: [dept, steps]
      properties:
        dept:
          type: string
          description: "Lead department ID for this branch (qualified with plan name prefix)."
        steps:
          type: array
          items: { $ref: '#/components/schemas/ExecutionStep' }
    SubWorkflowStep:
      type: object
      required: [name, plan]
      properties:
        node_id:     { type: string, description: "Stable BPMN element ID of the subProcess element." }
        name:        { type: string }
        plan:
          type: object
          properties:
            steps:
              type: array
              items: { $ref: '#/components/schemas/ExecutionStep' }
        error_paths:
          type: array
          items: { $ref: '#/components/schemas/ErrorPath' }
        timer_paths:
          type: array
          items: { $ref: '#/components/schemas/TimerPath' }
        message_paths:
          type: array
          items: { $ref: '#/components/schemas/MessagePath' }
          description: "From message boundary events attached to the subProcess."
    ErrorPath:
      type: object
      required: [short_circuit]
      properties:
        error_code:   { type: string, nullable: true, description: "BPMN error code that triggers this path. Empty means catch-all." }
        short_circuit: { type: boolean, description: "true = bypass remaining stages and route directly to target_dept." }
        target_dept:  { type: string, nullable: true, description: "Department to route to on this error path." }
    TimerPath:
      type: object
      required: [duration, interrupting]
      properties:
        duration:     { type: string, description: "ISO 8601 duration string." }
        interrupting: { type: boolean, description: "true = task cancelled on fire; false = escalation runs concurrently." }
        target_dept:  { type: string, nullable: true, description: "Department to route to on this timer path." }
    MessagePath:
      type: object
      required: [message_name, interrupting]
      properties:
        message_name: { type: string, description: "Resolved message name used for runtime correlation. May be empty if the source BPMN's message flow/messageRef could not be resolved (see MISSING_MESSAGE_DEFINITION warning)." }
        interrupting: { type: boolean, description: "true = host activity cancelled on fire; false = escalation runs concurrently." }
        target_dept:  { type: string, nullable: true, description: "Department to route to on this message path, if set." }
    CompiledPlan:
      type: object
      required: [name, departments, execution]
      properties:
        name:        { type: string }
        task_queue:  { type: string, description: "Temporal task queue. Omitted for ignored pools." }
        ignored:
          type: boolean
          description: "True when this pool was marked ignore=true in the BPMN. Consumers decide how to handle it."
        departments:
          type: array
          items: { $ref: '#/components/schemas/DepartmentDef' }
        execution:
          type: object
          properties:
            steps:
              type: array
              items: { $ref: '#/components/schemas/ExecutionStep' }
        visual_elements:
          type: array
          nullable: true
          description: "Visual-only BPMN elements (e.g. dataStoreReference) that carry no execution semantics. Informational only."
          items: { $ref: '#/components/schemas/VisualElementDef' }
      example:
        name: Tender Document Review Process
        task_queue: wf-queue-default
        departments:
          - id: Tender
            label: Tender
            stages:
              - { type: prep,    activity: PrepActivity,    role: bd-agent,       default_assignees: [018e1f2a-0000-7000-8000-000000000001], extras: { requires_comment: "false" } }
              - { type: review,  activity: ReviewActivity,  role: tender-manager, default_assignees: [018e1f2a-0000-7000-8000-000000000002], extras: { requires_comment: "true" } }
              - { type: approve, activity: ApproveActivity, role: tender-director, default_assignees: [018e1f2a-0000-7000-8000-000000000003], extras: { requires_comment: "true" } }
          - id: Engineering
            label: Engineering
            stages:
              - { type: prep,    activity: PrepActivity,    role: lead-engineer,  default_assignees: [018e1f2a-0000-7000-8000-000000000010], extras: { requires_comment: "false" }, boundary_timer: { duration: "PT48H", interrupting: false } }
        execution:
          steps:
            - sequential: [Tender, Engineering]
...
    ProblemDetails:
      type: object
      required: [type, title, status, detail, instance, code]
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem.
        code:
          type: string
          description: Service-specific error code for programmatic handling.
        invalid_params:
          type: array
          description: Details about specific fields that failed validation.
          items:
            type: object
            required: [name, reason]
            properties:
              name:
                type: string
              reason:
                type: string
              code:
                type: string
      example:
        type: "https://api.workflow.platform/errors/not-found"
        title: "Resource Not Found"
        status: 404
        detail: "Workflow not found or tenant boundary breached."
        instance: "/workflows/8fa88cde-824c-47bc-836b-665cd42c2222"
        code: "NOT_FOUND"
    VisualElementDef:
      type: object
      required: [kind, id]
      properties:
        kind: { type: string, description: "Element type tag (e.g. dataStoreReference)." }
        id:   { type: string, description: "BPMN element ID." }
        name: { type: string, nullable: true, description: "Display name, if set on the element." }
...
  responses:
    Unauthorized:
      description: Missing or invalid gateway identity headers.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
          example:
            type: "https://api.workflow.platform/errors/unauthorized"
            title: "Unauthorized Access"
            status: 401
            detail: "x-tenant-id header is missing or could not be resolved to a valid tenant."
            instance: "/workflows"
            code: "UNAUTHORIZED"
...
    NotFound:
      description: Resource not found or tenant boundary breached (security through obscurity).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
          example:
            type: "https://api.workflow.platform/errors/not-found"
            title: "Resource Not Found"
            status: 404
            detail: "Workflow not found or belongs to a different tenant."
            instance: "/workflows/8fa88cde-824c-47bc-836b-665cd42c2222"
            code: "NOT_FOUND"
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
          example:
            type: "https://api.workflow.platform/errors/internal-error"
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred. Please retry or contact support if the issue persists."
            instance: "/workflows"
            code: "INTERNAL_ERROR"
paths:
...
  /api/v1/workflows/{id}/versions/{version_id}:
    get:
      operationId: getVersion
      summary: Get version detail including BPMN XML and compiled DSL
      tags: [Versions]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TenantRoles'
        - $ref: '#/components/parameters/Departments'
        - $ref: '#/components/parameters/Plan'
        - $ref: '#/components/parameters/FeatureFlags'
        - $ref: '#/components/parameters/WorkflowId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        "200":
          description: Version detail.
          content:
            application/json:
              schema:
                type: object
                properties:
                  version_id:         { type: string, format: uuid }
                  workflow_id:        { type: string, format: uuid }
                  version_number:     { type: integer, nullable: true }
                  status:             { $ref: '#/components/schemas/VersionStatus' }
                  is_valid:           { type: boolean }
                  validation_errors:
                    type: array
                    items: { $ref: '#/components/schemas/ValidationError' }
                    nullable: true
                  bpmn_xml:           { type: string }
                  module_bpmn_xmls:   { type: array, items: { type: string }, description: "Called-process BPMNs attached to this version. Empty array if none." }
                  artifact_hash:      { type: string, nullable: true, description: "SHA-256 of the BPMN XML at publish time. Absent on DRAFT versions." }
                  compiled_plan_json: { $ref: '#/components/schemas/CompiledPlan' }
                  created_by_user_id: { type: string, format: uuid }
                  created_at:         { type: string, format: date-time }
                  published_at:       { type: string, format: date-time, nullable: true }
                  record_version:     { type: integer, format: int64, description: "Monotonic counter. Echo on PUT /draft requests." }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalError' }

  # 3.7
...